Fix const qualifier warnings with safe uintptr_t casts
Fixed 5 const-cast warnings:
1. FontResourceLoader: Changed FM_GetFontAssociation to return const FontAssocEntry**
2. DialogManager: Changed GetDialogManagerDispatchTable to return const void*
3. StandardFile: Added uintptr_t intermediate casts for FSSpec->name assignments
4. ResourceMgr: Added uintptr_t cast for const patterns_rsrc_data assignment

All casts use uintptr_t intermediate to avoid direct const-to-non-const casts.

Warnings: 588 → 583 (-5)

+++ src/ResourceMgr/ResourceMgr.c
@@ -247,7 +247,7 @@ void InitResourceManager(void) {
     extern const unsigned char patterns_rsrc_data[];
     extern const unsigned int patterns_rsrc_size;
 
-    gResMgr.resFiles[0].data = (UInt8*)patterns_rsrc_data;
+    gResMgr.resFiles[0].data = (UInt8*)(uintptr_t)patterns_rsrc_data;
     gResMgr.resFiles[0].dataSize = patterns_rsrc_size;
 
     /* Parse resource header and map */
